finegpt/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.3
faiss-cpu==1.8.0
sentence-transformers==3.2.0
numpy==1.26.4
pydantic==2.9.2
python-dotenv==1.0.1
